requirements.txt
flask
ultralytics
opencv-python
torch
pandas
fpdf2
openpyxl
